CORE-13353: Update docs with new settings changes (#336)
Updates all settings images that have or focus on the tabs as a means of navigating to the new updated UI portal now uses. To main things:

I did not focus on ALL settings pages, just the ones that included the out of date tabs. Any pages that now have the sidebar that previously did not show tabs has not been updated, as it's tangential to the point of that specific guide and thus extraneous.
This should wait to merge until next portal release so the docs match the live app